backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Session==0.5.0
PyMySQL==1.1.0
Werkzeug==2.3.7
